requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
openpyxl>=3.1.0
google-genai>=0.2.0
python-dotenv>=1.0.0
PyYAML>=6.0
streamlit-authenticator>=0.3.2
reportlab>=4.0.0
streamlit-option-menu>=0.3.6
kaleido>=0.2.1
schedule>=1.2.0
